+++ core/src/test/java/org/apache/mahout/vectorizer/collocations/llr/LLRReducerTest.java
@@ -59,7 +59,7 @@ public final class LLRReducerTest extends MahoutTestCase {
     ll        = EasyMock.createMock(LLCallback.class);
     cl        = new LLCallback() {
       @Override
-      public double logLikelihoodRatio(int k11, int k12, int k21, int k22) {
+      public double logLikelihoodRatio(long k11, long k12, long k21, long k22) {
         log.info("k11:{} k12:{} k21:{} k22:{}", new Object[] {k11, k12, k21, k22});
         return LogLikelihood.logLikelihoodRatio(k11, k12, k21, k22);
       }
